requirements.txt
beautifulsoup4
requests
faiss-cpu
sentence-transformers
llama-cpp-python
gradio
